Refactor CreateVoiceActivity scope and dependencies
