+++ app/src/main/java/org/ole/planet/myplanet/lite/util/ApplicationScope.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.lite
+package org.ole.planet.myplanet.lite.util
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
